+++ Toggly.FeatureManagement.Nuxt/nuxt-toggly-core/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.9.1
+
+2026-09-03
+
+### Changed
+- Raise `@ops-ai/toggly-eval` floor to `^2.0.3` [OPS-874].
+
 ## 1.9.0
 
 2026-09-03